+++ apps/web/package.json
@@ -76,7 +76,6 @@
     "remark-toc": "^9.0.0",
     "resend": "catalog:",
     "schema-dts": "^1.1.5",
-    "server-only": "^0.0.1",
     "tailwind-merge": "^3.0.2",
     "tailwindcss-animate": "^1.0.7",
     "zod": "catalog:",
